[NodeTypeResolver] decouple ClassLikeTypeResolver
